The app plays in the world read from eFootball: game_world.db, dt200 clubs, real PIDs
Ruling 2026-09-13: the MVP uses build/game_world.db only (master.db stays on disk, unread) and is
limited to what is native to eFootball. A career is started with a dt200 club, edited, played in
eFootball, and its stats come back through the Rust host.

- One resolver (WorldFiles) replaces five copies of "walk up to build/master.db" in the loader,
  catalog, vault, market and seeder launch. Play Match and the seeder are told the world
  explicitly; both used to fall back to master.db on their own.
- game_world.py builds from dt200 only (Play Match can only render a club with a dt200 slot),
  writes the picker's game_catalog.json, and fills an internal strength from derive_strength.py's
  per-position fit (held-out: 94.5% exact, 100% within one point; the app is to show stars).
- Career ids follow the world (SessionBands, meta career_player_band): a game-built world keeps
  1,578 real players in 20M-700M, where careers used to copy players and the Market hid them.
  The seeder's clean slate would also have deleted the band key itself, and the next career would
  then have wiped those real players — world facts are now kept like preferences are.
- Career copies carry the real PID (base_pid) and the game's playstyles and skills. Play Match
  resolves a club by its base_team_id and a player by base_pid before any name matching, and the
  stats importer links an export by the PID the match was played under. The collaborator's open
  issue — career fixtures never linked — closes here. BasePid is a long (card PIDs pass 2^44).
- Play Match no longer needs the editor's export: the in-possession style table comes from
  player_layouts.json, where it was proven 100% and is not position-scoped.

Verified: an Arsenal career seeds into game_world.db (44 clubs, 2 tiers, 933 fixtures, 1,439 copies
all with real PIDs and playstyles, the 1,078 real 20M-700M players untouched); Arsenal v Liverpool
compiles with both slots and all 75 players resolved by id; build clean, 208/208 tests.

+++ src/ML.Data/Records.cs
@@ -54,8 +54,11 @@ public sealed record PlayerRow
     // Liverpool and Arsenal, and every screen built on Session.Squad() — Tactics first — died
     // before it existed. The write side (academy prospects, sample data) already fits.
     public long GamePid { get; init; }
-    public int? BasePid { get; init; }
-    public int? DonorPid { get; init; }
+    // LONG for the same reason: base_pid is a real eFootball PID, and the current game's club cards
+    // carry PIDs past 2^44. As an int? the first such row would throw out of Repository, exactly as
+    // GamePid did on Tactics.
+    public long? BasePid { get; init; }
+    public long? DonorPid { get; init; }
     public bool IsCustom { get; init; }
     public string Name { get; init; } = "";
     public string? ShortName { get; init; }
